chore: move stdout probe to separate dir
